requirements.txt
Flask>=2.0
python-dotenv
google-cloud-firestore
Flask-WTF
Flask-Login
gunicorn
pytest
